+++ packages/fuselage-hooks/tests/useExclusiveBooleanProps.spec.js
@@ -1,26 +1,26 @@
-import { testHook } from '../.jest/helpers';
+import { runHooks } from '../.jest/helpers';
 import { useExclusiveBooleanProps } from '../src';
 
 describe('useExclusiveBooleanProps hook', () => {
   it('allows an empty set of props', () => {
-    testHook(() => useExclusiveBooleanProps({}));
+    runHooks(() => useExclusiveBooleanProps({}));
   });
 
   it('allows only false-valued props', () => {
-    testHook(() => useExclusiveBooleanProps({ a: false, b: false, c: false }));
+    runHooks(() => useExclusiveBooleanProps({ a: false, b: false, c: false }));
   });
 
   it('allows one true-valued prop', () => {
-    testHook(() => useExclusiveBooleanProps({ a: true }));
+    runHooks(() => useExclusiveBooleanProps({ a: true }));
   });
 
   it('allows one true-valued prop among false-valued ones', () => {
-    testHook(() => useExclusiveBooleanProps({ a: true, b: false, c: false }));
+    runHooks(() => useExclusiveBooleanProps({ a: true, b: false, c: false }));
   });
 
   it('denies two true-valued props', () => {
     expect(() => {
-      testHook(() => useExclusiveBooleanProps({ a: true, b: true }));
+      runHooks(() => useExclusiveBooleanProps({ a: true, b: true }));
     }).toThrow();
   });
 });