+++ packages/fuselage-ui-kit/src/hooks.js
@@ -1,4 +1,4 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext, useState, useCallback } from 'react';
 import {
   BLOCK_CONTEXT,
 } from '@rocket.chat/ui-kit';
@@ -14,26 +14,28 @@ export const kitContext = React.createContext(defaultContext);
 
 export const useBlockContext = ({ blockId, actionId, appId, initialValue }, context) => {
   const { action, appId: appIdFromContext, viewId, state, errors, values = {} } = useContext(kitContext);
-  const { value = initialValue } = values[actionId] || {};
-  // const [value, setValue] = useState(initialValue);
+  const { value: _value = initialValue } = values[actionId] || {};
+  const [value, setValue] = useState(_value);
   const [loading, setLoading] = useState(false);
 
   const error = errors && actionId && errors[actionId];
 
+  const actionFunction = useCallback(async ({ target: { value } }) => {
+    setLoading(true);
+    await action({ blockId, appId: appId || appIdFromContext, actionId, value, viewId });
+    setLoading(false);
+  }, [actionId, blockId]);
+
+  const stateFunction = useCallback(async ({ target: { value } }) => {
+    setValue(value);
+    await state({ blockId, appId, actionId, value });
+  }, [actionId, blockId]);
+
   if ([BLOCK_CONTEXT.SECTION, BLOCK_CONTEXT.ACTION].includes(context)) {
-    return [{ loading, setLoading, error }, async ({ target: { value } }) => {
-      setLoading(true);
-      await action({ blockId, appId: appId || appIdFromContext, actionId, value, viewId });
-      setLoading(false);
-    }];
+    return [{ loading, setLoading, error }, actionFunction];
   }
 
-  return [{ loading, setLoading, value, error }, async ({ target: { value } }) => {
-    // setValue(value);
-    setLoading(true);
-    await state({ blockId, appId, actionId, value });
-    setLoading(false);
-  }];
+  return [{ loading, setLoading, value, error }, stateFunction];
 };
 
 export const getStyle = (style) => {